Timestamp: paso a uint64_t y cambio la función get_timestamp()

+++ kernel/interfaz.h
@@ -23,7 +23,7 @@ void imprimir_metricas();
 
 
 //Prototipos
-void kernel_insert(int socket_memoria, char* nombre_tabla, uint16_t key, char* value, int epoch);
+void kernel_insert(int socket_memoria, char* nombre_tabla, uint16_t key, char* value, uint64_t epoch);
 void kernel_select(int socket_memoria, char *tabla, uint16_t key);
 void kernel_create(int socket_memoria, char* nombre_tabla, char* tipo_consistencia, int numero_particiones, int compaction_time);
 void kernel_describe(int socket_memoria, char* nombre_tabla);
