+++ memoria/interfaz.h
@@ -8,7 +8,7 @@
 
 
 char *memoria_select(char *tabla, uint16_t key);
-void memoria_insert(char *tabla, uint16_t key, char *value, int timestamp);
+void memoria_insert(char *tabla, uint16_t key, char *value, uint64_t timestamp);
 void memoria_create(char *tabla, char *consistencia, int particiones, int compaction_time);
 void memoria_describe(char *tabla);
 void memoria_drop(char *tabla);
