+++ shared_library/estructuras_compartidas.h
@@ -22,7 +22,7 @@ typedef struct {
 	char* nombre_tabla;
 	uint16_t key;
 	char* value;
-	int epoch;
+	uint64_t epoch;
 } t_request_insert;
 //Variables
 
@@ -50,7 +50,7 @@ typedef struct {
 void* serializar_request_select(char*tabla, uint16_t key);
 t_request_select *deserializar_request_select(t_prot_mensaje *msje);
 
-void* serializar_request_insert(char* nombre_tabla, uint16_t key, char* value, int epoch);
+void* serializar_request_insert(char* nombre_tabla, uint16_t key, char* value, uint64_t epoch);
 t_request_insert *deserializar_request_insert(t_prot_mensaje *msje);
 
 void* serializar_request_create(char* nombre_tabla, char* tipo_consistencia, int numero_particiones, int compaction_time);
